I made slide show using margin -%
The project used buttons to apply a slide-changing css, not an image slide that automatically changes over time. This time, the Css gave the effect of moving the picture using margin, but it was regrettable that it had to move in one direction only, and that it changed at a fast pace when moving the first slide in the fifth slide.